install zod and validate job input

+++ controllers/jobController.js
@@ -1,4 +1,6 @@
 import Job from "../models/jobModel.js";
+import { errorHandler } from "../utils/errorHandler.js";
+import { JobSchema } from "../validationSchema/jobsSchema.js";
 
 const getAllJobs = async (req, res, next) => {
   try {
@@ -13,7 +15,15 @@ const getAllJobs = async (req, res, next) => {
 const getJob = async (req, res, next) => {
   try {
     const { id } = req.params;
+    if (!id) {
+      console.log("Invalid id");
+      return next(errorHandler(400, "Something went wrong"));
+    }
     const job = await Job.findById(id);
+    if (!job) {
+      console.log("Job not found");
+      return next(errorHandler(400, "Something went wrong"));
+    }
     res.status(200).json(job);
   } catch (error) {
     console.log(error);
@@ -23,11 +33,12 @@ const getJob = async (req, res, next) => {
 
 const createJob = async (req, res, next) => {
   try {
-    const { company, position } = req.body;
-    const newJob = await Job.create({
-      company,
-      position,
-    });
+    const result = JobSchema.safeParse(req.body);
+    if (!result.success) {
+      console.log(result.error);
+      return next(errorHandler(400, "Invalid data"));
+    }
+    const newJob = await Job.create(req.body);
     res.status(201).json(newJob);
   } catch (error) {
     console.log(error);
@@ -37,6 +48,11 @@ const createJob = async (req, res, next) => {
 const deleteJob = async (req, res, next) => {
   try {
     const { id } = req.params;
+    const job = await Job.findById(id);
+    if (!job) {
+      console.log("Job not found");
+      return next(errorHandler(400, "Something went wrong"));
+    }
     await Job.deleteOne({ _id: id });
     res.status(200).json({ msg: "successfully deleted" });
   } catch (error) {
@@ -47,6 +63,16 @@ const deleteJob = async (req, res, next) => {
 const updateJob = async (req, res, next) => {
   try {
     const { id } = req.params;
+    const job = await Job.findById(id);
+    if (!job) {
+      console.log("Job not found");
+      return next(errorHandler(400, "Something went wrong"));
+    }
+    const result = JobSchema.safeParse(req.body);
+    if (!result.success) {
+      console.log(result.error);
+      return next(errorHandler(400, "Invalid data"));
+    }
     const newJob = await Job.findByIdAndUpdate(id, req.body, { new: true });
     res.status(201).json(newJob);
   } catch (error) {
